Changes per code review from PR
